submissions-team/andy-chen/adv_deployment/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
Pillow==10.0.0
python-jose==3.3.0
pydantic==2.4.2
python-dotenv==1.0.0
opencv-python==4.8.1.78 